# docs/swagger.yaml
definitions:
...
  models.Product:
    properties:
      category_id:
        type: integer
      price:
        type: integer
      stock:
        type: integer
      title:
        type: string
      transaction_history:
        items:
          $ref: '#/definitions/models.TransactionHistory'
        type: array
    type: object
  models.TransactionHistory:
    properties:
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: integer
      quantity:
        type: integer
      total_price:
        type: integer
      user_id:
        type: integer
    type: object
  models.User:
    properties:
      balance:
        type: integer
      email:
        type: string
      full_name:
        type: string
      password:
        type: string
      role:
        type: string
      transaction_history:
        items:
          $ref: '#/definitions/models.TransactionHistory'
        type: array
    type: object
info:
  contact: {}
paths:
...
  /users/register:
    post:
      parameters:
      - description: Balance
        in: body
        name: email
        required: true
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "201":
          description: User registered successfully
          schema:
            $ref: '#/definitions/models.User'
      summary: Updates the user's balance
swagger: "2.0"